Разнёс валидацию пропсов по отдельным файлам
