+++ src/components/offer-list/offer-list.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import {PropTypes} from 'prop-types';
+import {offersPropValid} from './offer-card/offer-card.prop';
 
 import OfferCard from './offer-card/offer-card';
-import {offersPropValid} from '../../props-valid/props-valid';
 
 
 const OfferList = ({offers, path}) => (
