+++ src/components/offer-property/review-list/review-item/review-item.prop.js
@@ -1,0 +1,14 @@
+import PropTypes from 'prop-types';
+
+export const reviewsPropValid = {
+  comment: PropTypes.string.isRequired,
+  date: PropTypes.string.isRequired,
+  id: PropTypes.number.isRequired,
+  rating: PropTypes.number.isRequired,
+  user: PropTypes.shape({
+    avatarUrl: PropTypes.string.isRequired,
+    id: PropTypes.number.isRequired,
+    isPro: PropTypes.bool.isRequired,
+    name: PropTypes.string.isRequired,
+  }).isRequired,
+};
